Moved the message log components to a subfolder
