+++ bootstrapper/bootstrapper.ts
@@ -2,6 +2,6 @@ import { services } from 'typescript-angular-utilities';
 import __timezone = services.timezone;
 
 import './bootstrapApp';
-import './messageLogBootstrapper';
+import './messageLog/messageLogBootstrapper';
 import './cardContainerBootstrapper';
 import './text';
