+++ bootstrapper/messageLog/messageLogBootstrapper.js
